requirements.txt
flask-mongoalchemy
flask-principal
flask-wtf
flask-script
flask-babel
Flask-Cache
Flask-SQLAlchemy
Flask-Mail
Flask-Uploads
flask-markdown